requirements.txt
absl-py==2.1.0
astunparse==1.6.3
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
dbus-python==1.2.18
distro==1.7.0
flatbuffers==24.3.25
gast==0.5.4
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.10.0
idna==3.6
keras==2.15.0
libclang==18.1.1
Markdown==3.6
MarkupSafe==2.1.5
ml-dtypes==0.2.0
numpy==1.26.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==24.0
pillow==10.2.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
scipy==1.12.0
six==1.16.0
ssh-import-id==5.11
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
wrapt==1.14.1
opencv-python==4.9.0.80
